Fix sql init script do second Criteria from task 	modified:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DAO/CustomerRepository.java 	new file:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DAO/ProductRepository.java 	new file:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DAO/ReceiptRepository.java 	modified:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DTO/Input/Criteria.java 	modified:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/Services/CliActions.java 	modified:   src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/Services/Search.java 	modified:   src/main/resources/application.properties 	modified:   src/main/resources/data.sql

+++ src/main/java/com/github/thecoolersuptelov/aikamsoftjavajunior/DTO/Input/Criteria.java
@@ -21,7 +21,7 @@ public class Criteria implements Serializable {
     @JsonProperty("productName")
     public String productName;
     @JsonProperty("minTimes")
-    public int minTimes;
+    public long minTimes;
     @JsonProperty("minExpenses")
     public int minExpenses;
     @JsonProperty("maxExpenses")
